+++ CHANGELOG.md
@@ -51,4 +51,5 @@
 
 ## 1.5.1
 
-- [Lib] Fix: Step methods resolution from library file with multiple classes. Fixes #51
+- [Lib] Fix: Step methods resolution from library file with multiple classes. Fixes [#51](https://github.com/PiotrFLEURY/pickled_cucumber/issues/51)
+- [Lib] Fix: Only parametrized steps resolution in scenarios. Fixes [#28](https://github.com/PiotrFLEURY/pickled_cucumber/issues/28)